requirements.txt
opencv-python
opencv-python-headless
numpy
imutils
pyserial
cvzone
